Refactor: Tích hợp Vulkan Memory Allocator (VMA)
Thay thế toàn bộ logic cấp phát bộ nhớ thủ công (vkAllocateMemory, vkBindMemory, findMemoryTypeIndex) bằng thư viện VMA.

- Thêm `VmaAllocator` vào `VulkanContext` và quản lý vòng đời của nó.
- Refactor `VulkanBuffer` để sử dụng `vmaCreateBuffer`/`vmaDestroyBuffer`:
  - Thay thế `VkDeviceMemory` bằng `VmaAllocation`.
  - Đơn giản hóa `UploadData` để tự động map bộ nhớ (`CPU_TO_GPU`) hoặc dùng staging (`GPU_ONLY`) dựa trên `VmaMemoryUsage`.
- Refactor `VulkanImage` để sử dụng `vmaCreateImage`/`vmaDestroyImage`.
- Cập nhật các lớp gọi (main, MeshManager) để sử dụng cờ `VmaMemoryUsage` thay vì boolean.

+++ VulkanLearn2/Scene/Model.cpp
@@ -13,7 +13,7 @@ Model::Model(const std::string& modelFilePath, MeshManager* meshManager, Texture
 	// 2. Dùng MeshManager để tạo các đối tượng Mesh từ dữ liệu thô.
 	//    Hàm này sẽ đưa dữ liệu vertex/index vào các buffer chung và trả về các struct Mesh
 	//    chứa thông tin về vị trí và kích thước của dữ liệu đó trong buffer.
-	m_Handles.meshes = meshManager->createMeshFromMeshData(modelData.meshData.data(), modelData.meshData.size());
+	m_Handles.meshes = meshManager->createMeshFromMeshData(modelData.meshData.data(),static_cast<uint32_t>(modelData.meshData.size()));
 
 	// 3. Với mỗi mesh, tải texture tương ứng thông qua TextureManager.
 	for (int i = 0; i < m_Handles.meshes.size(); i++)
